Got light combos working
Twix was nice too (Tesco was completely empty, imagine that, they even had a sign in the lobby saying all customers limited to 3 of any item unless it was a multibuy). Got a pineapple and white chocolate meese.
May need to work on timings, the second attack is a mess. It also seems to occurr every time. Turning the likelihood of this down may be better for a more interesting boss.