Started limiting the time the sword is activated
On the one hand, it may be a little harsh, and make the game slightly more difficult, but on the other, it also means there's less of a chance to be hit by a lingering hitbox, so swings and roundabouts.
This was a very big problem with Heavy 2, where the sword does an all over the world dance, and the stab was pretty bad too.